custom token count updated, Agency UI added
